+++ Web/Memorial/Core/Domain/Site.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Web;
 
 namespace Memorial.Core.Domain
 {
@@ -24,7 +22,7 @@ namespace Memorial.Core.Domain
             Catalogs = new HashSet<Catalog>();
         }
 
-        public byte Id { get; set; }
+        public int Id { get; set; }
 
         public string Name { get; set; }
