+++ Web/Memorial/Core/Domain/Space.cs
@@ -22,7 +22,7 @@ namespace Memorial.Core.Domain
 
         public Site Site { get; set; }
 
-        public byte SiteId { get; set; }
+        public int SiteId { get; set; }
 
         public string ColorCode { get; set; }
