+++ Web/Memorial/Core/Dtos/ColumbariumCentreDto.cs
@@ -10,7 +10,7 @@
 
         public SiteDto SiteDto { get; set; }
 
-        public byte SiteDtoId { get; set; }
+        public int SiteDtoId { get; set; }
 
     }
 }
