+++ Web/Memorial/Core/Dtos/CremationDto.cs
@@ -12,6 +12,6 @@ namespace Memorial.Core.Dtos
 
         public SiteDto SiteDto { get; set; }
 
-        public byte SiteDtoId { get; set; }
+        public int SiteDtoId { get; set; }
     }
 }
